requirements.txt
pyyaml==6.0.2
python-dotenv==1.0.1
tzdata==2025.1
pydantic>=2.6,<3
pydantic-settings>=2.1,<3
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2

# IQ Option API (upstream repo; PyPI package is outdated)
websocket-client==0.56.0
iqoptionapi @ https://github.com/iqoptionapi/iqoptionapi/archive/refs/heads/master.zip
PySocks>=1.7.1,<2


streamlit>=1.33,<2

matplotlib>=3.9,<4
reportlab>=4.2,<5
